rewrite rtdb in typescript, using apiv2
